Make LayerIdx and FrameIdx strong types

+++ src/label widget.hpp
@@ -32,7 +32,7 @@ LabelWidget *makeLabel(QWidget *parent, const char (&text)[Size]) {
 
 template <size_t Size>
 LabelWidget *makeLabel(QWidget *parent, const int chars, const char (&text)[Size]) {
-  assert(Size - 1 <= chars);
+  assert(static_cast<int>(Size) - 1 <= chars);
   return new LabelWidget{parent, textBoxRect(chars), text};
 }
